+++ .g8/MultipleQuestionsDecimalPage/app/models/$className$.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class $className$ ($field1Name$: Double, $field2Name$: Double)
+case class $className$ ($field1Name$: Long, $field2Name$: Long)
 
 object $className$ {
   implicit val format = Json.format[$className$]
